pyproject.toml
[project]
name = "webpage-sniffer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "azure-functions>=1.24.0",
    "pytz>=2025.2",
]
